[Finishing course] (tasks 13, 19)
- Now the mark for course is float to avoid problems with rounding.
- Created methods to get average max and success mark for course based on success and max mark of each lesson.
- Created class for manual actions on startup: OnStartUpRunner.

 - Changed the logic of "task 13" - we calculate course mark based on average lessons' mark.
 - Now on course finish we mark course as finished, count mark of course for each participating, set if user passed each lesson based on lesson's mark

+++ src/main/java/com/nazar/grynko/learningcourses/dto/usertocourse/UserToCourseDto.java
@@ -17,7 +17,7 @@ import java.io.Serializable;
 public class UserToCourseDto implements Serializable {
 
     private Long id;
-    private Integer mark;
+    private Float mark;
     @JsonProperty("isPassed")
     private Boolean isPassed;
     private Long userId;
